Reorganize test files in internal/cli/
Move orphaned test files into their corresponding source file test files:

- context_test.go → root_test.go (tests ExecuteContext)
- flags_test.go → init_test.go (tests interactiveWizard/quickSetup quiet mode)
- logging_test.go → root_test.go (tests verbose/quiet output behavior)
- spinner_test.go → root_test.go (tests outcomeToCompletion)
- json_types_test.go → split into auth_test.go, config_test.go,
  key_test.go, init_test.go, cache_test.go (each JSON test goes with
  the command it exercises)

Shared test helpers consolidated:
- reloadConfig() from helpers_test.go → testutil_test.go
- newVerboseContext/newQuietContext/newDefaultContext from
  logging_test.go → testutil_test.go

logger_test.go kept as-is (has matching logger.go source file).

Completes phase 3 of FIXIT_PLAN.md.
